Reogranization of file structure
