--- treeppl/examples/crbd/io.mc
@@ -1,10 +1,0 @@
-include "tree.mc"
--- This file is instead of IO
--- The values should be supplied from "outside" somehow
--- tree is provided by tree-instance.mc
-let k_l = 1.0
-let t_l = 1.0
-let k_m = 1.0
-let t_m = 0.5
-
-let observation = Node {left = Node {left = Node {left = Node {left = Leaf {age = 0.0}, right = Leaf {age = 0.0}, age = 5.635787971}, right = Node {left = Leaf {age = 0.0}, right = Node {left = Leaf {age = 0.0}, right = Node {left = Leaf {age = 0.0}, right = Leaf {age = 0.0}, age = 4.788021775}, age = 7.595901077}, age = 9.436625313}, age = 12.344087935000001}, right = Node {left = Node {left = Leaf {age = 0.0}, right = Node {left = Node {left = Leaf {age = 0.0}, right = Leaf {age = 0.0}, age = 3.934203877}, right = Node {left = Node {left = Leaf {age = 0.0}, right = Leaf {age = 0.0}, age = 3.151799953}, right = Node {left = Leaf {age = 0.0}, right = Leaf {age = 0.0}, age = 5.054547857}, age = 6.284896356999999}, age = 7.815689970999999}, age = 10.32243059}, right = Node {left = Leaf {age = 0.0}, right = Node {left = Node {left = Leaf {age = 0.0}, right = Node {left = Leaf {age = 0.0}, right = Leaf {age = 0.0}, age = 1.519406055}, age = 4.987038163}, right = Node {left = Leaf {age = 0.0}, right = Node {left = Node {left = Leaf {age = 0.0}, right = Leaf {age = 0.0}, age = 0.6302632958}, right = Node {left = Leaf {age = 0.0}, right = Node {left = Leaf {age = 0.0}, right = Leaf {age = 0.0}, age = 1.962579854}, age = 3.732932004}, age = 5.5933070698}, age = 6.096453021}, age = 8.265483252}, age = 10.86835485}, age = 12.551924091}, age = 13.472886809}, right = Node {left = Node {left = Node {left = Leaf {age = 0.0}, right = Node {left = Leaf {age = 0.0}, right = Leaf {age = 0.0}, age = 4.534421013}, age = 12.46869821}, right = Node {left = Leaf {age = 0.0}, right = Node {left = Leaf {age = 0.0}, right = Node {left = Leaf {age = 0.0}, right = Node {left = Leaf {age = 0.0}, right = Leaf {age = 0.0}, age = 6.306427821}, age = 9.40050129}, age = 13.85876825}, age = 20.68766993}, age = 22.82622451}, right = Node {left = Leaf {age = 0.0}, right = Node {left = Node {left = Leaf {age = 0.0}, right = Node {left = Leaf {age = 0.0}, right = Node {left = Leaf {age = 0.0}, right = Node {left = Leaf {age = 0.0}, right = Leaf {age = 0.0}, age = 4.220057646}, age = 8.451051062}, age = 11.54072627}, age = 15.28839572}, right = Node {left = Node {left = Node {left = Leaf {age = 0.0}, right = Leaf {age = 0.0}, age = 8.614086751}, right = Node {left = Leaf {age = 0.0}, right = Node {left = Leaf {age = 0.0}, right = Node {left = Leaf {age = 0.0}, right = Node {left = Node {left = Leaf {age = 0.0}, right = Leaf {age = 0.0}, age = 0.9841688636}, right = Node {left = Leaf {age = 0.0}, right = Leaf {age = 0.0}, age = 1.04896206}, age = 1.7140599232}, age = 3.786162534}, age = 8.788450495}, age = 11.05846217}, age = 15.008504768}, right = Node {left = Node {left = Leaf {age = 0.0}, right = Leaf {age = 0.0}, age = 11.15685875}, right = Node {left = Leaf {age = 0.0}, right = Node {left = Leaf {age = 0.0}, right = Node {left = Leaf {age = 0.0}, right = Node {left = Leaf {age = 0.0}, right = Node {left = Leaf {age = 0.0}, right = Leaf {age = 0.0}, age = 1.900561313}, age = 3.100150132}, age = 6.043650727}, age = 12.38252513}, age = 12.61785812}, age = 15.396725736}, age = 16.828404506}, age = 20.368109703000002}, age = 23.74299959}, age = 32.145876657}, age = 34.940139089}
