--- treeppl/examples/flip/data.mc
@@ -1,1 +1,0 @@
-let p = 0.5
