--- treeppl/examples/flip/flip.tppl
@@ -1,7 +1,0 @@
-/* Comment */
-// another comment
-model function flip(p: Real): Bool {
-    e ~ Bernoulli(p)
-    return e
-}
-
