+++ treeppl/models/flip/data.mc
@@ -1,0 +1,1 @@
+let p = 0.5
