+++ treeppl/models/flip/flip-io.mc
@@ -1,0 +1,4 @@
+-- This file is instead of IO
+-- The values should be supplied from "outside" somehow
+-- tree is provided by tree-instance.mc
+let p = 0.5
